% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logconcave.R
\name{bounds.logconc.internal}
\alias{bounds.logconc.internal}
\title{Computes upper identification interval under the assumption that F is log-concave.}
\usage{
bounds.logconc.internal(X, sampling.ratio = 5, xmin = NULL, xmax = NULL,
  buckets = 1000, alpha = 1/sqrt(length(X)))
}
\arguments{
\item{X}{The observed data.}

\item{sampling.ratio}{Bound on the sampling weights gamma.}

\item{xmin}{Used to construct histogram representation.}

\item{xmax}{Used to construct histogram representation.}

\item{buckets}{Used to construct histogram representation.}

\item{alpha}{Significance level used for KS bounds.}
}
\value{
mu.bound The upper bound for mu(x).

Fhat Unweighted empirical CDF of the data.

xvals Points at which Fhat is evaluated.

Fhat.upper KS bound for Fhat.

Lhat.upper Lhat function from paper.

Weighted version of Fhat that maximizes mu, subject to log-concavity
}
\description{
Computes upper identification interval under the assumption that F is log-concave.
}

