% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\title{Main function. Computes identification interval.}
\usage{
bounds(X, sampling.ratio = 5, constraint = c("none", "logconcave",
  "symmetric", "gaussian"), alpha = 1/length(X), xmin = min(X),
  xmax = max(X), buckets = 1000)
}
\arguments{
\item{X}{The observed data.}

\item{sampling.ratio}{Bound on the sampling weights gamma.}

\item{constraint}{Optional constraint on the shape of F.}

\item{alpha}{Significance level used for KS bounds.}

\item{xmin}{Used to construct histogram representation.}

\item{xmax}{Used to construct histogram representation.}

\item{buckets}{Used to construct histogram representation.}
}
\value{
mu.interval The identification interval for mu(x).

lower.bound.internal Internal data used to create lower bound.

upper.bound.internal Internal data used to create upper bound.
}
\description{
Main function. Computes identification interval.
}

