% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds.upper.internal}
\alias{bounds.upper.internal}
\title{Computes identification intervals under the assumption that F is log-concave.}
\usage{
bounds.upper.internal(X, alpha = 1/length(X), sampling.ratio = 5,
  xmin = NULL, xmax = NULL, buckets = 1000)
}
\arguments{
\item{X}{The observed data.}

\item{alpha}{Significance level used for KS bounds.}

\item{sampling.ratio}{Bound on the sampling weights gamma.}

\item{xmin}{Used to construct histogram representation.}

\item{xmax}{Used to construct histogram representation.}

\item{buckets}{Used to construct histogram representation.}
}
\value{
Fhat Unweighted empirical CDF of the data.

xvals Points at which Fhat is evaluated.

Fhat.upper KS bound for Fhat.

Lhat.upper Lhat function from paper.

Weighted version of Fhat that maximizes mu, subject to log-concavity

Weighted version of Fhat that maximizes mu, without log-concavity constraint
}
\description{
Computes identification intervals under the assumption that F is log-concave.
}

