% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.ks.threshold}
\alias{get.ks.threshold}
\title{Find (one-sided) length of Kolmogorov-Smirnov bounds for a CDF}
\usage{
get.ks.threshold(n, alpha = 0.05)
}
\arguments{
\item{n}{The number of samples used to compute Fhat.}

\item{alpha}{The significance level.}
}
\value{
The KS threshold
}
\description{
Find (one-sided) length of Kolmogorov-Smirnov bounds for a CDF
}

