% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{get.ks.upper}
\alias{get.ks.upper}
\title{Find level-alpha Kolmogorov-Smirnov bounds for a CDF F}
\usage{
get.ks.upper(Fhat, n, alpha = 0.05)
}
\arguments{
\item{Fhat}{The empirical CDF.}

\item{n}{The number of samples used to compute Fhat.}

\item{alpha}{The significance level.}
}
\value{
The ''upper'' KS bound (i.e., the one that lies below Fhat).
}
\description{
Find level-alpha Kolmogorov-Smirnov bounds for a CDF F
}

