% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{get.logconc.upper}
\alias{get.logconc.upper}
\title{Find the smallest log-concave upper envelope for a function}
\usage{
get.logconc.upper(H)
}
\arguments{
\item{H}{The function H(x), specified at equally spaced x-values.}
}
\value{
Lhat.upper The log-concave upper bound.
}
\description{
Find the smallest log-concave upper envelope for a function
}

