% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logconcave.R
\name{get.logconc.upper.scan}
\alias{get.logconc.upper.scan}
\title{Finds the "Uhat" function from the paper, i.e., the pointwise minimum of all log-concave
function that lie above some weighted version of Fhat.upper.}
\usage{
get.logconc.upper.scan(Fhat.upper, sampling.ratio)
}
\arguments{
\item{Fhat.upper}{The unweighted function.}

\item{sampling.ratio}{Bound on the sampling weights.}
}
\value{
Lhat.upper The pointwise minimal bound.
}
\description{
Finds the "Uhat" function from the paper, i.e., the pointwise minimum of all log-concave
function that lie above some weighted version of Fhat.upper.
}

