% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hajek.constrained.R
\name{hajek.constrained}
\alias{hajek.constrained}
\title{Computational workhorse function. Maximizes the value of hat{mu} among all Hajek
ratio estimators that satisfy the following properties:
- The ratio of the sampling weights gamma is bounded
- The resulting weighted CDF lies "above" lower.bound (as a function)}
\usage{
hajek.constrained(Fhat, xvals, lower.bound, sampling.ratio = 5)

hajek.constrained(Fhat, xvals, lower.bound, sampling.ratio = 5)
}
\arguments{
\item{Fhat}{The raw, unweighted, empirical CDF.}

\item{xvals}{The points at which Fhat is specified.}

\item{lower.bound}{The lower bound for the Hajek-weighted empirical CDF.}

\item{sampling.ratio}{The bound on the sampling ratio.}

\item{Fhat}{The raw, unweighted, empirical CDF.}

\item{xvals}{The points at which Fhat is specified.}

\item{center}{The center of symmetry.}

\item{delta}{The tolerance parameter.}

\item{sampling.ratio}{The bound on the sampling ratio.}
}
\value{
Fhat.weighted A weighted version of Fhat that maximizes hat{mu}

Fhat.weighted A weighted version of Fhat that maximizes hat{mu}
}
\description{
Computational workhorse function. Maximizes the value of hat{mu} among all Hajek
ratio estimators that satisfy the following properties:
- The ratio of the sampling weights gamma is bounded
- The resulting weighted CDF lies "above" lower.bound (as a function)

Computational workhorse function. Maximizes the value of hat{mu} among all Hajek
ratio estimators that satisfy the following properties:
- The ratio of the sampling weights gamma is bounded
- The resulting CDF is symmetric to within tolerance delta.
}

