% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hajek.constrained.R
\name{hajek.constrained.symmetric}
\alias{hajek.constrained.symmetric}
\title{Computational workhorse function. Maximizes the value of hat{mu} among all Hajek
ratio estimators that satisfy the following properties:
- The ratio of the sampling weights gamma is bounded
- The resulting CDF is symmetric to within tolerance delta.}
\usage{
hajek.constrained.symmetric(Fhat, xvals, center, delta, sampling.ratio = 5)
}
\arguments{
\item{Fhat}{The raw, unweighted, empirical CDF.}

\item{xvals}{The points at which Fhat is specified.}

\item{center}{The center of symmetry.}

\item{delta}{The tolerance parameter.}

\item{sampling.ratio}{The bound on the sampling ratio.}
}
\value{
Fhat.weighted A weighted version of Fhat that maximizes hat{mu}
}
\description{
Computational workhorse function. Maximizes the value of hat{mu} among all Hajek
ratio estimators that satisfy the following properties:
- The ratio of the sampling weights gamma is bounded
- The resulting CDF is symmetric to within tolerance delta.
}

